+++ packages/transaction-controller/src/types.ts
@@ -1800,6 +1800,9 @@ export type GasFeeToken = {
   /** Estimated gas limit required for fee transfer. */
   gasTransfer?: Hex;
 
+  /** Whether the gas costs are sponsored meaning a transfer is not required. */
+  isSponsored?: boolean;
+
   /** The corresponding maxFeePerGas this token fee would equal. */
   maxFeePerGas: Hex;
 
@@ -1817,6 +1820,7 @@ export type GasFeeToken = {
 
   /** Address of the token contract. */
   tokenAddress: Hex;
+
 };
 
 /** Request to check if atomic batch is supported for an account. */